+++ gd-glm/src/containers/Layout/Footer.jsx
@@ -5,7 +5,6 @@ import {
     Grid,
     makeStyles
 } from '@material-ui/core';
-
 import LogoNCSA from '../../images/logo_ncsa.png';
 import LogoSeaGrant from '../../images/logo_seagrant.png';
 import LogoGLRI from '../../images/logo_glri.png';
@@ -107,7 +106,7 @@ const Footer = () => {
                         target="_blank"
                         rel="noopener noreferrer"
                     >
-                        Geodashboard v.3.3.0
+                        Geodashboard v.{process.env.VERSION}
                     </a>
                 </Grid>
             </Grid>