+++ gd-geostreaming/CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## UNRELEASED
+## [3.5.0] - 2020-10-01
 
 ### Added
 - Sensor detail view
   [GEOD-1341](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1341)
+
+### Changed
+- Use encoded variable for sensor names in the url
+  [GEOD-1347](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1347)