+++ gd-glm/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.5.0] - 2020-10-01
+### Changed
+- Decrease the zoom level when a source is selected
+  [GLM-127](https://opensource.ncsa.illinois.edu/jira/browse/GLM-127)
+- Fixed a typo in the source section of the config file
+  [GLM-131](https://opensource.ncsa.illinois.edu/jira/browse/GLM-131)
+
+
 ## [3.4.0] - 2020-08-27
 ### Fixed
 - Fix the missing logo in the header at non-root url paths