+++ gd-gltg/CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## UNRELEASED
+## [3.5.0] - 2020-10-01
 
 ### Changed
 - Improved the components and config relations between `gd-gltg` and `gd-geostreaming`
   [GEOD-1341](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1341)
+- Get version number in footer from package.json
+  [GLGVO-666](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-666)
 
 ## [3.3.0] - 2020-06-18